score_service/requirements.txt
fastapi
uvicorn
psycopg2-binary
numpy
requests
pydantic 